Move `OntologyTypeVertexId` to `subgraph` module
