+++ apps/hash-graph/lib/graph/src/subgraph.rs
@@ -1,1 +1,3 @@
+pub mod identifier;
+
 pub use crate::shared::subgraph::*;
