+++ apps/hash-graph/tests/integration/postgres/lib.rs
@@ -21,7 +21,7 @@ use graph::{
         knowledge::EntityId,
         ontology::OntologyTypeVersion,
         time::{DecisionTime, LimitedTemporalBound, TemporalBound, Timestamp},
-        GraphElementVertexId, OntologyTypeVertexId,
+        GraphElementVertexId,
     },
     knowledge::{
         Entity, EntityLinkOrder, EntityMetadata, EntityProperties, EntityQueryPath, EntityUuid,
@@ -41,6 +41,7 @@ use graph::{
     },
     subgraph::{
         edges::GraphResolveDepths,
+        identifier::OntologyTypeVertexId,
         query::StructuralQuery,
         temporal_axes::{
             PinnedTemporalAxisUnresolved, QueryTemporalAxesUnresolved,
